+++ src/commons_codec/model.py
@@ -6,7 +6,7 @@ from enum import auto
 if sys.version_info >= (3, 11):
     from enum import StrEnum
 else:
-    from backports.strenum import StrEnum
+    from backports.strenum import StrEnum  # pragma: no cover
 
 from attrs import define