+++ tests/zyp/conftest.py
@@ -1,0 +1,6 @@
+# ruff: noqa: E402
+import pytest
+
+jmespath = pytest.importorskip("jmespath")
+jsonpointer = pytest.importorskip("jsonpointer")
+transon = pytest.importorskip("transon")